Перенесены аннотации Swagger из ProjectService в интерфейс контроллера
